requirements.txt
Django>=5.0.2
python-dotenv>=1.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-allauth>=0.58.2
django-celery-results>=2.5.1
django-celery-beat>=2.5.0
cloudinary>=1.37.1
django-cloudinary-storage>=0.3.0
paypalrestsdk>=1.13.1
razorpay>=2.8.4
Pillow>=10.2.0
PyPDF2>=3.0.1
python-docx>=1.1.0
numpy>=1.26.4
scikit-learn>=1.4.0
pandas>=2.2.1
tensorflow>=2.15.0
textblob>=0.17.1 